requirements.txt
mysql-connector-python==8.0.31
redis==4.2.0